perf: File directory adjustment
