+++ apps/ops/api/__init__.py
@@ -2,7 +2,6 @@
 #
 from .adhoc import *
 from .celery import *
-from .inventory import *
 from .job import *
 from .playbook import *
 from .variable import *
